chore: move fees to the account level
we will need to be able to set the fees for each account.
