+++ src/services/mongoose/wallets.ts
@@ -82,8 +82,6 @@ export const WalletsRepository = (): IWalletsRepository => {
 
 const resultToWallet = (result: UserType): Wallet => {
   const walletId = result.walletId as WalletId
-  const depositFeeRatio = result.depositFeeRatio as DepositFeeRatio
-  const withdrawFee = result.withdrawFee as WithdrawFee
 
   const onChainAddressIdentifiers = result.onchain.map(({ pubkey, address }) => {
     return {
@@ -95,15 +93,11 @@ const resultToWallet = (result: UserType): Wallet => {
 
   return {
     id: walletId,
-    depositFeeRatio,
-    withdrawFee,
     onChainAddressIdentifiers,
     onChainAddresses,
   }
 }
 const projection = {
   walletId: 1,
-  depositFeeRatio: 1,
-  withdrawFee: 1,
   onchain: 1,
 }
